email_func_app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Pillow==10.1.0
python-dotenv==1.0.0
email-validator==2.1.0.post1
PyMySQL==1.1.0  # Python MySQL client library
cryptography==41.0.5  # Required for PyMySQL
